pyproject.toml
[project]
name = "openelectionsdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "email-validator>=2.3.0",
    "pydantic==2.12.5",
    "pytest>=9.0.2",
    "python-levenshtein>=0.27.3",
]
